Remove sandpack examples
